[CA] Moved listers to util package
